api/requirements.txt
Django>=4
djangorestframework>=3.12.0
djangorestframework-simplejwt>=5.2.2
drf-spectacular>=0.15.1
Pillow>=9.4.0
psycopg2-binary>=2.8.6